add -w option
Providing passwords via command line is insecure.  Adding an option to read
the password from the TTY or stdin.  When reading from a TTY, getpass() is
used so that the password is not displayed as it is entered.  Getpass()
also displays a prompt, since TTY usage is typically interactive.  When
reading from something other than a TTY (e.g. from a pipe or input
redirection), a plain read from stdin is done (without getpass(); without
any promting).  This allows the user to pipe the password from some type of
password management/entry tool to registry.py.